requirements.txt
unsloth
transformers
trl
peft
accelerate
bitsandbytes
datasets
scikit-learn
pandas
pyyaml
tqdm
openai
huggingface_hub
